requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.20.0
Flask==3.1.0
flask-cors==6.0.1
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.41
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
marshmallow==3.26.1
modelscope==1.31.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
networkx==3.4.2
numpy==2.2.6
openai==1.109.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==25.2
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-settings==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
zstandard==0.25.0